requirements.txt
Django==4.2
python-decouple==3.8
stripe==12.4.0
pillow==11.3.0
Werkzeug==3.1.3
PyOpenSSL==25.1.0
django-extensions==4.1
django-sslserver==0.22